requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.3
tzdata==2024.2

